finance_analyst/pyproject.toml
[project]
name = "finance-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2026.1.4",
    "firecrawl-py>=4.12.0",
    "google-adk>=1.21.0",
    "python-dotenv>=1.2.1",
]
